@@ -1,41 +1,26 @@
-﻿using System.Numerics;
+﻿using MikuMikuMethods.Pmm.ElementState;
 
-namespace MikuMikuMethods.PMM.Frame
+namespace MikuMikuMethods.Pmm.Frame;
+
+public class PmmAccessoryFrame : PmmAccessoryState, IPmmFrame
 {
-    /// <summary>
-    /// アクセサリフレーム情報
-    /// </summary>
-    public class PmmAccessoryFrame : PmmFrame
+    public int Frame { get; set; }
+    public bool IsSelected { get; set; }
+
+    public new PmmAccessoryFrame DeepCopy() => new()
     {
-        /// <summary>
-        /// <para>上位7bit : 半透明度</para>
-        /// <para>最下位1bit : 表示/非表示</para>
-        /// </summary>
-        public byte OpacityAndVisible { get; set; }
-        /// <summary>
-        /// <para>親モデルのインデックス</para>
-        /// <para>-1 なら親なし</para>
-        /// </summary>
-        public int ParentModelIndex { get; set; }
-        /// <summary>
-        /// 親ボーンのインデックス
-        /// </summary>
-        public int ParentBoneIndex { get; set; }
-        /// <summary>
-        /// 位置
-        /// </summary>
-        public Vector3 Position { get; set; }
-        /// <summary>
-        /// 回転
-        /// </summary>
-        public Vector3 Rotation { get; set; }
-        /// <summary>
-        /// 拡縮
-        /// </summary>
-        public float Scale { get; set; }
-        /// <summary>
-        /// 影のOn/Off
-        /// </summary>
-        public bool EnableShadow { get; set; }
-    }
+        Frame = Frame,
+        IsSelected = IsSelected,
+        EnableShadow = EnableShadow,
+        ParentModel = ParentModel,
+        ParentBone = ParentBone,
+        Position = Position,
+        Rotation = Rotation,
+        Scale = Scale,
+        TransAndVisible = TransAndVisible,
+    };
+
+    IPmmFrame IPmmFrame.DeepCopy() => DeepCopy();
+
+    public override string ToString() => Frame.ToString();
 }
