@@ -1,552 +1,787 @@
-﻿using MikuMikuMethods.Extension;
-using MikuMikuMethods.PMM.Frame;
-using System.IO;
+using MikuMikuMethods.Common;
+using MikuMikuMethods.Extension;
+using MikuMikuMethods.Pmm.ElementState;
+using MikuMikuMethods.Pmm.Frame;
 
-namespace MikuMikuMethods.PMM.IO
+namespace MikuMikuMethods.Pmm.IO;
+
+public static class PmmFileReader
 {
-    internal static class PmmFileReader
+    public delegate void OnSectionChangeEventHandler(DataSection section);
+    public static event OnSectionChangeEventHandler? OnChangeSection;
+
+    private static Dictionary<PmmModelConfigFrame, Dictionary<PmmBone, (int ModelID, int BoneID)>> OuterParentRelation { get; set; } = new();
+    private static Dictionary<PmmModelConfigState, Dictionary<PmmBone, (int ModelID, int BoneID)>> OuterParentRelationCurrent { get; set; } = new();
+    public static DataSection Current
     {
-        public static void Read(string filePath, PolygonMovieMaker pmm)
+        get => current; set
         {
-            using (FileStream stream = new(filePath, FileMode.Open))
-            using (BinaryReader reader = new(stream, Encoding.ShiftJIS))
+            current = value;
+            OnChangeSection?.Invoke(value);
+        }
+    }
+
+    private static DataSection current = new("", null, "");
+
+    public static void Read(string filePath, PolygonMovieMaker pmm)
+    {
+        try
+        {
+            using FileStream file = new(filePath, FileMode.Open);
+            using BinaryReader reader = new(file, Encoding.ShiftJIS);
+            Read(reader, pmm);
+        }
+        catch (Exception)
+        {
+            throw;
+        }
+    }
+
+
+    public static void Read(BinaryReader reader, PolygonMovieMaker pmm)
+    {
+        try
+        {
+            Current = new("Header", null, "Header section of the PMM file.");
+            pmm.Version = reader.ReadString(30, Encoding.ShiftJIS, '\0');
+            if (pmm.Version != "Polygon Movie maker 0002") throw new InvalidDataException("This is not PMM file.");
+            pmm.OutputResolution = new(reader.ReadInt32(), reader.ReadInt32());
+
+            pmm.EditorState.Width = reader.ReadInt32();
+            pmm.Camera.Current.ViewAngle = (int)reader.ReadSingle();
+            pmm.EditorState.IsCameraMode = reader.ReadBoolean();
+
+            Current = new("PanelOpeningStatus", null, "Section of the opening status of various operation panels.");
+            // パネル開閉状態の読み込み
+            pmm.PanelPane.DoesOpenCameraPanel = reader.ReadBoolean();
+            pmm.PanelPane.DoesOpenLightPanel = reader.ReadBoolean();
+            pmm.PanelPane.DoesOpenAccessaryPanel = reader.ReadBoolean();
+            pmm.PanelPane.DoesOpenBonePanel = reader.ReadBoolean();
+            pmm.PanelPane.DoesOpenMorphPanel = reader.ReadBoolean();
+            pmm.PanelPane.DoesOpenSelfShadowPanel = reader.ReadBoolean();
+
+            Current = new("HeaderOfModels", null, "Section of selected model index and total number of models.");
+            // モデル読み込み
+            var selectedModelIndex = reader.ReadByte();
+            var modelCount = reader.ReadByte();
+            var modelOrderDictionary = new Dictionary<PmmModel, (byte RenderOrder, byte CalculateOrder)>();
+            for (int i = 0; i < modelCount; i++)
             {
-                pmm.Version = reader.ReadString(30, Encoding.ShiftJIS, '\0');
-                pmm.Output = new(reader.ReadInt32(), reader.ReadInt32());
+                Current = new("Model", i, $"Section of {DataSection.GetOrdinal(i)} model data.");
+                (var model, var renderOrder, var calculateOrder) = ReadModel(reader);
+                pmm.Models.Add(model);
+                modelOrderDictionary.Add(model, ((byte RenderOrder, byte CalculateOrder))(renderOrder - 1, calculateOrder - 1));
+            }
 
-                ReadViewState(reader, pmm.EditorState);
+            Current = new($"ModelRelationSolving", null, $"The section that resolves the relations of the selected model and the render/calculate order. In this section, no reading is done, only calculation.");
+            pmm.EditorState.SelectedModel = selectedModelIndex < pmm.Models.Count ? pmm.Models[selectedModelIndex] : null;
+            foreach (var model in pmm.Models)
+            {
+                // 順序系プロパティはモデルの追加後に設定する
+                var (renderOrder, calculateOrder) = modelOrderDictionary[model];
+                pmm.SetRenderOrder(model, renderOrder);
+                pmm.SetCalculateOrder(model, calculateOrder);
+            }
 
-                var modelCount = reader.ReadByte();
-                for (int i = 0; i < modelCount; i++)
-                    pmm.Models.Add(ReadModel(reader));
-
-                ReadCamera(reader, pmm.Camera);
-                ReadLight(reader, pmm.Light);
-
-                ReadAccessoryState(reader, pmm.EditorState);
-                var accessoryCount = reader.ReadByte();
-                // アクセサリ名一覧
-                // 名前は各アクセサリ領域にも書いてあり、齟齬が出ることは基本無いらしいので読み飛ばす
-                _ = reader.ReadBytes(accessoryCount * 100);
-                for (int i = 0; i < accessoryCount; i++)
-                    pmm.Accessories.Add(ReadAccessory(reader));
-
-                ReadFrameState(reader, pmm.EditorState);
-                ReadPlayConfig(reader, pmm.PlayConfig);
-                ReadMediaConfig(reader, pmm.MediaConfig);
-                ReadDrawConfig(reader, pmm.ViewConfig);
-                ReadGravity(reader, pmm.Gravity);
-                ReadSelfShadow(reader, pmm.SelfShadow);
-                ReadColorConfig(reader, pmm.ViewConfig);
-                ReadUncomittedFollowingState(reader, pmm.Camera);
-                //謎の行列は読み飛ばす
-                _ = reader.ReadBytes(64);
-                ReadViewFollowing(reader, pmm.ViewConfig);
-                pmm.Unknown = new PmmUnknown { TruthValue = reader.ReadBoolean() };
-                ReadGroundPhysics(reader, pmm.ViewConfig);
-                ReadFrameLocation(reader, pmm.ViewConfig);
-
-                // バージョンによってはここで終わりの可能性がある
-                if (reader.BaseStream.Position >= reader.BaseStream.Length)
-                    return;
-
-                pmm.EditorState.ExistRangeSelectionTargetSection = reader.ReadBoolean();
-                // 普通はないが範囲選択対象セクションが無いとされていれば終了
-                if (!pmm.EditorState.ExistRangeSelectionTargetSection)
-                    return;
-
-                // モデル設定値が欠落している場合もあるっぽい？ので確認処理を挟む
-                for (byte i = 0; i < pmm.Models.Count; i++)
+            // 外部親の関係解決
+            Current = new("OuterParentSolving", null, $"The section that resolves the outer parent relation. In this section, no reading is done, only calculation.");
+            foreach (var frame in OuterParentRelation)
+            {
+                foreach (var relation in frame.Value)
                 {
-                    if (reader.BaseStream.Position >= reader.BaseStream.Length)
-                        break;
-
-                    pmm.EditorState.RangeSelectionTargetIndices.Add(ReadRangeSelectionTargetIndex(reader, pmm));
-                    //pmm.EditorState.RangeSelectionTargetIndices.Add((reader.ReadByte(), reader.ReadInt32()));
+                    var opModel = relation.Value.ModelID < 0 ? null : pmm.Models[relation.Value.ModelID];
+                    frame.Key.OuterParent.Add(relation.Key, new()
+                    {
+                        ParentModel = opModel,
+                        ParentBone = opModel?.Bones[relation.Value.BoneID]
+                    });
                 }
             }
-        }
 
-        private static void ReadViewState(BinaryReader reader, PmmEditorState editorState)
-        {
-            editorState.KeyframeEditorWidth = reader.ReadInt32();
+            foreach (var state in OuterParentRelationCurrent)
+            {
+                foreach (var relation in state.Value)
+                {
+                    var opModel = relation.Value.ModelID < 0 ? null : pmm.Models[relation.Value.ModelID];
+                    state.Key.OuterParent[relation.Key].ParentModel = opModel;
+                    state.Key.OuterParent[relation.Key].ParentBone = opModel?.Bones[relation.Value.BoneID];
+                }
+            }
 
-            editorState.CurrentViewAngle = reader.ReadSingle();
+            ReadCamera(reader, pmm);
+            ReadLight(reader, pmm.Light);
 
-            editorState.IsCameraMode = reader.ReadBoolean();
+            Current = new("HeaderOfAccessories", null, "Section of selected accessory index and total number of accessories.");
+            var selectedAccessoryIndex = reader.ReadByte();
+            pmm.EditorState.VerticalScrollOfAccessory = reader.ReadInt32();
 
-            editorState.DoesOpenCameraPanel = reader.ReadBoolean();
-            editorState.DoesOpenLightPanel = reader.ReadBoolean();
-            editorState.DoesOpenAccessaryPanel = reader.ReadBoolean();
-            editorState.DoesOpenBonePanel = reader.ReadBoolean();
-            editorState.DoesOpenMorphPanel = reader.ReadBoolean();
-            editorState.DoesOpenSelfShadowPanel = reader.ReadBoolean();
+            // アクセサリ読み込み
+            var accessoryOrderDictionary = new Dictionary<PmmAccessory, byte>();
+            var accessoryCount = reader.ReadByte();
+            // アクセサリ名一覧
+            // 名前は各アクセサリ領域にも書いてあり、齟齬が出ることは基本無いらしいので読み飛ばす
+            _ = reader.ReadBytes(accessoryCount * 100);
+            for (int i = 0; i < accessoryCount; i++)
+            {
+                Current = new($"Accessory", i, $"The section of {DataSection.GetOrdinal(i)} accessory data.");
+                (PmmAccessory accessory, byte renderOrder) = ReadAccessory(reader, pmm);
+                pmm.Accessories.Add(accessory);
+                accessoryOrderDictionary.Add(accessory, renderOrder);
+            }
 
-            editorState.SelectedModelIndex = reader.ReadByte();
-        }
+            Current = new($"AccessoryRelationSolving", null, $"The section that resolves the relations of the selected accessory and the render order. In this section, no reading is done, only calculation.");
+            pmm.EditorState.SelectedAccessory = selectedAccessoryIndex < pmm.Accessories.Count ? pmm.Accessories[selectedAccessoryIndex] : null;
+            foreach (var acs in pmm.Accessories)
+            {
+                pmm.SetRenderOrder(acs, accessoryOrderDictionary[acs]);
+            }
 
-        private static PmmModel ReadModel(BinaryReader reader)
-        {
-            var model = new PmmModel();
+            Current = new($"KeyFrameEditorCurrentTarget", null, $"Sections for the current keyframe editor editing target and scroll amount.");
+            // フレーム編集画面の状態読み込み
+            pmm.EditorState.CurrentFrame = reader.ReadInt32();
+            pmm.EditorState.HorizontalScroll = reader.ReadInt32();
+            pmm.EditorState.HorizontalScrollLength = reader.ReadInt32();
+            pmm.EditorState.SelectedBoneOperation = (PmmEditorState.BoneOperation)reader.ReadInt32();
 
-            model.Index = reader.ReadByte();
+            Current = new($"Media", null, $"The sections of play config ,background media, render config.");
+            // 再生関連設定の読込
+            pmm.PlayConfig.CameraTrackingTarget = (PmmPlayConfig.TrackingTarget)reader.ReadByte();
 
-            model.Name = reader.ReadString();
-            model.NameEn = reader.ReadString();
-            model.Path = reader.ReadString(256, Encoding.ShiftJIS, '\0');
+            pmm.PlayConfig.EnableRepeat = reader.ReadBoolean();
+            pmm.PlayConfig.EnableMoveCurrentFrameToPlayStopping = reader.ReadBoolean();
+            pmm.PlayConfig.EnableStartFromCurrentFrame = reader.ReadBoolean();
 
-            // キーフレームエディタの行数
-            // 表情枠数から求まるので破棄
+            pmm.PlayConfig.PlayStartFrame = reader.ReadInt32();
+            pmm.PlayConfig.PlayStopFrame = reader.ReadInt32();
+
+            // 背景メディア設定の読込
+            var existsBgm = reader.ReadBoolean();
+            var bgmPath = reader.ReadString(256, Encoding.ShiftJIS, '\0');
+            pmm.BackGround.Audio = existsBgm ? bgmPath : null;
+
+            pmm.BackGround.VideoOffset = new(reader.ReadInt32(), reader.ReadInt32());
+            pmm.BackGround.VideoScale = reader.ReadSingle();
+            var bgvPath = reader.ReadString(256, Encoding.ShiftJIS, '\0');
+            var existsBgv = reader.ReadInt32() == 0b01000000;
+            pmm.BackGround.Video = existsBgv ? bgvPath : null;
+
+            pmm.BackGround.ImageOffset = new(reader.ReadInt32(), reader.ReadInt32());
+            pmm.BackGround.ImageScale = reader.ReadSingle();
+            var bgiPath = reader.ReadString(256, Encoding.ShiftJIS, '\0');
+            var existsBgi = reader.ReadBoolean();
+            pmm.BackGround.Image = existsBgi ? bgiPath : null;
+
+            // 描画設定の読込
+            pmm.RenderConfig.InfomationVisible = reader.ReadBoolean();
+            pmm.RenderConfig.AxisVisible = reader.ReadBoolean();
+            pmm.RenderConfig.EnableGrandShadow = reader.ReadBoolean();
+
+            pmm.RenderConfig.FPSLimit = (PmmRenderConfig.FPSLimitValue)(int)reader.ReadSingle();
+            pmm.RenderConfig.ScreenCaptureMode = (PmmRenderConfig.ScreenCaptureModeType)reader.ReadInt32();
+
+            pmm.RenderConfig.PostDrawingAccessoryStartIndex = reader.ReadInt32();
+            pmm.RenderConfig.GroundShadowBrightness = reader.ReadSingle();
+            pmm.RenderConfig.EnableTransparentGroundShadow = reader.ReadBoolean();
+
+            ReadPhysics(reader, pmm.Physics);
+            ReadSelfShadow(reader, pmm.SelfShadow);
+
+            Current = new("FollowingSettings", null, "Camera following settings and more");
+            pmm.RenderConfig.EdgeColor = System.Drawing.Color.FromArgb(reader.ReadInt32(), reader.ReadInt32(), reader.ReadInt32());
+            pmm.BackGround.IsBlack = reader.ReadBoolean();
+
+            int currentCameraFollowingModelIndex = reader.ReadInt32();
+            int currentCameraFollowingBoneIndex = reader.ReadInt32();
+            pmm.Camera.Current.FollowingModel = currentCameraFollowingModelIndex > 0 ? pmm.Models[currentCameraFollowingModelIndex] : null;
+            pmm.Camera.Current.FollowingBone = pmm.Camera.Current.FollowingModel?.Bones[currentCameraFollowingBoneIndex];
+
+            // 意図不明な謎の行列
+            _ = reader.ReadBytes(64);
+
+            pmm.PlayConfig.EnableFollowCamera = reader.ReadBoolean();
+
+            // 意図不明な謎の値
             _ = reader.ReadByte();
 
-            var boneCount = reader.ReadInt32();
-            for (int i = 0; i < boneCount; i++)
+            pmm.Physics.EnableGroundPhysics = reader.ReadBoolean();
+            pmm.RenderConfig.JumpFrameLocation = reader.ReadInt32();
+
+            try
             {
-                model.BoneNames.Add(reader.ReadString());
-            }
+                // バージョン 9.24 より前ならここで終了のため、 EndOfStreamException が投げられる
+                // 9.24 なら後続要素が存在するかの Boolean 値が読める
+                var existSelectorChoiseSection = reader.ReadBoolean();
 
-            var morphCount = reader.ReadInt32();
-            for (int i = 0; i < morphCount; i++)
-            {
-                model.MorphNames.Add(reader.ReadString());
-            }
-
-            var ikCount = reader.ReadInt32();
-            for (int i = 0; i < ikCount; i++)
-            {
-                model.IKBoneIndices.Add(reader.ReadInt32());
-            }
-
-            var parentableBoneCount = reader.ReadInt32();
-            for (int i = 0; i < parentableBoneCount; i++)
-            {
-                model.ParentableBoneIndices.Add(reader.ReadInt32());
-            }
-
-            model.RenderConfig.RenderOrder = reader.ReadByte();
-
-            model.Uncomitted.Visible = reader.ReadBoolean();
-            model.SelectedBoneIndex = reader.ReadInt32();
-            model.SelectedMorphIndices = (reader.ReadInt32(), reader.ReadInt32(), reader.ReadInt32(), reader.ReadInt32());
-
-            var nodeCount = reader.ReadByte();
-            for (int i = 0; i < nodeCount; i++)
-            {
-                model.FrameEditor.DoesOpenNode.Add(reader.ReadBoolean());
-            }
-
-            model.FrameEditor.VerticalScrollState = reader.ReadInt32();
-            model.FrameEditor.LastFrame = reader.ReadInt32();
-
-            for (int i = 0; i < boneCount; i++)
-            {
-                model.InitialBoneFrames.Add(ReadBoneFrame(reader, null));
-            }
-
-            var boneFrameCount = reader.ReadInt32();
-            for (int i = 0; i < boneFrameCount; i++)
-            {
-                model.BoneFrames.Add(ReadBoneFrame(reader, reader.ReadInt32()));
-            }
-
-            for (int i = 0; i < morphCount; i++)
-            {
-                model.InitialMorphFrames.Add(ReadMorphFrame(reader, null));
-            }
-
-            var morphFrameCount = reader.ReadInt32();
-            for (int i = 0; i < morphFrameCount; i++)
-            {
-                model.MorphFrames.Add(ReadMorphFrame(reader, reader.ReadInt32()));
-            }
-
-            model.InitialConfigFrame = ReadConfigFrame(reader, null, ikCount, parentableBoneCount);
-
-            var configFrameCount = reader.ReadInt32();
-            for (int i = 0; i < configFrameCount; i++)
-            {
-                model.ConfigFrames.Add(ReadConfigFrame(reader, reader.ReadInt32(), ikCount, parentableBoneCount));
-            }
-
-            for (int i = 0; i < boneCount; i++)
-            {
-                TemporaryBoneState uncomittedBone = new();
-                uncomittedBone.Offset = reader.ReadVector3();
-                uncomittedBone.Rotate = reader.ReadQuaternion();
-                uncomittedBone.IsThis = reader.ReadBoolean();
-                uncomittedBone.EnablePhysic = reader.ReadBoolean();
-                uncomittedBone.RowIsSelected = reader.ReadBoolean();
-
-                model.Uncomitted.Bones.Add(uncomittedBone);
-            }
-
-            for (int i = 0; i < morphCount; i++)
-            {
-                model.Uncomitted.MorphWeights.Add(reader.ReadSingle());
-            }
-
-            for (int i = 0; i < ikCount; i++)
-            {
-                model.Uncomitted.IKEnable.Add(reader.ReadBoolean());
-            }
-
-            for (int i = 0; i < parentableBoneCount; i++)
-            {
-                TemporaryParentSetting uncomittedParentSetting = new();
-                uncomittedParentSetting.StartFrame = reader.ReadInt32();
-                uncomittedParentSetting.EndFrame = reader.ReadInt32();
-                uncomittedParentSetting.ModelIndex = reader.ReadInt32();
-                uncomittedParentSetting.BoneIndex = reader.ReadInt32();
-
-                model.Uncomitted.ParentSettings.Add(uncomittedParentSetting);
-            }
-
-            model.RenderConfig.EnableAlphaBlend = reader.ReadBoolean();
-            model.RenderConfig.EdgeWidth = reader.ReadSingle();
-            model.RenderConfig.EnableSelfShadow = reader.ReadBoolean();
-            model.RenderConfig.CalculateOrder = reader.ReadByte();
-
-            return model;
-
-            PmmBoneFrame ReadBoneFrame(BinaryReader reader, int? index)
-            {
-                var frame = new PmmBoneFrame();
-
-                frame.Index = index;
-
-                frame.Frame = reader.ReadInt32();
-                frame.PreviousFrameIndex = reader.ReadInt32();
-                frame.NextFrameIndex = reader.ReadInt32();
-
-                frame.InterpolationCurces[InterpolationItem.XPosition].FromBytes(reader.ReadBytes(4));
-                frame.InterpolationCurces[InterpolationItem.YPosition].FromBytes(reader.ReadBytes(4));
-                frame.InterpolationCurces[InterpolationItem.ZPosition].FromBytes(reader.ReadBytes(4));
-                frame.InterpolationCurces[InterpolationItem.Rotation].FromBytes(reader.ReadBytes(4));
-
-                frame.Offset = reader.ReadVector3();
-                frame.Rotation = reader.ReadQuaternion();
-                frame.IsSelected = reader.ReadBoolean();
-                frame.EnablePhysic = !reader.ReadBoolean();
-
-                return frame;
-            }
-
-            PmmMorphFrame ReadMorphFrame(BinaryReader reader, int? index) => new()
-            {
-                Index = index,
-
-                Frame = reader.ReadInt32(),
-                PreviousFrameIndex = reader.ReadInt32(),
-                NextFrameIndex = reader.ReadInt32(),
-
-                Ratio = reader.ReadSingle(),
-
-                IsSelected = reader.ReadBoolean(),
-            };
-
-            PmmConfigFrame ReadConfigFrame(BinaryReader reader, int? index, int ikCount, int parentCount)
-            {
-                var frame = new PmmConfigFrame();
-
-                frame.Index = index;
-
-                frame.Frame = reader.ReadInt32();
-                frame.PreviousFrameIndex = reader.ReadInt32();
-                frame.NextFrameIndex = reader.ReadInt32();
-                frame.Visible = reader.ReadBoolean();
-
-                for (int i = 0; i < ikCount; i++)
+                // 存在しなければここ以降の情報は無意味なので読み飛ばす
+                // が MMD はそんな値は吐かないと思われる
+                if (existSelectorChoiseSection)
                 {
-                    frame.EnableIK.Add(reader.ReadBoolean());
+                    // 範囲選択セレクタの読込
+                    for (int i = 0; i < modelCount; i++)
+                    {
+                        Current = new("RangeSelector", i, $"The section of {DataSection.GetOrdinal(i)} range selection target.");
+                        pmm.Models[reader.ReadByte()].SpecificEditorState.RangeSelector = new(reader.ReadInt32());
+                    }
                 }
+            }
+            catch (EndOfStreamException)
+            {
+                // このセクションは途中でファイルが終わってても構わないので
+                // ストリームの終わり例外なら来ても何もしなくてよい
+            }
+        }
+        catch (Exception ex)
+        {
+            IOException exception = new($"Failed to read PMM file. This exception occurred in {Current.Name}. See Data[\"Section\"] property, that type is {Current.GetType().Name}, of this exception for details on where exceptions are occurred.", ex);
+            exception.Data.Add("Section", Current);
+            throw exception;
+        }
+        finally
+        {
+            OuterParentRelation = new();
+            OuterParentRelationCurrent = new();
+        }
+    }
 
-                for (int i = 0; i < parentCount; i++)
-                {
-                    frame.ParentSettings.Add((reader.ReadInt32(), reader.ReadInt32()));
-                }
+    private static void ReadCamera(BinaryReader reader, PolygonMovieMaker pmm)
+    {
 
-                frame.IsSelected = reader.ReadBoolean();
+        var camera = pmm.Camera;
 
-                return frame;
+        Current = new("CameraFrame", null, $"The section of initial camera frame data.");
+        camera.Frames.Add(ReadCameraFrame(reader, pmm, true));
+
+        var cameraFrameCount = reader.ReadInt32();
+        for (int i = 0; i < cameraFrameCount; i++)
+        {
+            Current = Current with { Index = i, Description = $"The section of {DataSection.GetOrdinal(i)} camera frame data." };
+            camera.Frames.Add(ReadCameraFrame(reader, pmm));
+        }
+
+        Current = new("CurrentCamera", null, $"The section of camera settings.");
+        camera.Current.EyePosition = reader.ReadVector3();
+        camera.Current.TargetPosition = reader.ReadVector3();
+        camera.Current.Rotation = reader.ReadVector3();
+        camera.Current.DisablePerspective = reader.ReadBoolean();
+    }
+    private static PmmCameraFrame ReadCameraFrame(BinaryReader reader, PolygonMovieMaker pmm, bool isInitial = false)
+    {
+        // リストの添え字で管理できるため不要なフレームインデックスを破棄
+        if (!isInitial) _ = reader.ReadInt32();
+
+        var frame = new PmmCameraFrame();
+
+        frame.Frame = reader.ReadInt32();
+
+        // 所属が確実にわかるので pre/next ID から探索してやる必要性がないため破棄
+        _ = reader.ReadInt32();
+        _ = reader.ReadInt32();
+
+        frame.Distance = reader.ReadSingle();
+        frame.EyePosition = reader.ReadVector3();
+        frame.Rotation = reader.ReadVector3();
+
+        int followingModelIndex = reader.ReadInt32();
+        int followingBoneIndex = reader.ReadInt32();
+
+        frame.FollowingModel = followingModelIndex < 0 ? null : pmm.Models[followingModelIndex];
+        frame.FollowingBone = frame.FollowingModel?.Bones[followingBoneIndex];
+
+        frame.InterpolationCurves[InterpolationItem.XPosition].FromBytes(reader.ReadBytes(4));
+        frame.InterpolationCurves[InterpolationItem.YPosition].FromBytes(reader.ReadBytes(4));
+        frame.InterpolationCurves[InterpolationItem.ZPosition].FromBytes(reader.ReadBytes(4));
+        frame.InterpolationCurves[InterpolationItem.Rotation].FromBytes(reader.ReadBytes(4));
+        frame.InterpolationCurves[InterpolationItem.Distance].FromBytes(reader.ReadBytes(4));
+        frame.InterpolationCurves[InterpolationItem.ViewAngle].FromBytes(reader.ReadBytes(4));
+
+        frame.DisablePerspective = reader.ReadBoolean();
+        frame.ViewAngle = reader.ReadInt32();
+
+        frame.IsSelected = reader.ReadBoolean();
+
+        return frame;
+    }
+
+    private static void ReadLight(BinaryReader reader, PmmLight light)
+    {
+        Current = new("LightFrame", null, $"The section of initial light frame data.");
+        light.Frames.Add(ReadLightFrame(reader, true));
+
+        var frameCount = reader.ReadInt32();
+        for (int i = 0; i < frameCount; i++)
+        {
+            Current = Current with { Index = i, Description = $"The section of {DataSection.GetOrdinal(i)} light frame data." };
+            light.Frames.Add(ReadLightFrame(reader));
+        }
+
+        Current = new("CurrentLight", null, $"The section of light settings.");
+        light.Current.Color = reader.ReadSingleRGB();
+        light.Current.Position = reader.ReadVector3();
+    }
+    private static PmmLightFrame ReadLightFrame(BinaryReader reader, bool isInitial = false)
+    {
+        // リストの添え字で管理できるため不要なフレームインデックスを破棄
+        if (!isInitial) _ = reader.ReadInt32();
+
+        var frame = new PmmLightFrame();
+
+        frame.Frame = reader.ReadInt32();
+        // 所属が確実にわかるので pre/next ID から探索してやる必要性がないため破棄
+        _ = reader.ReadInt32();
+        _ = reader.ReadInt32();
+
+        frame.Color = reader.ReadSingleRGB();
+        frame.Position = reader.ReadVector3();
+
+        frame.IsSelected = reader.ReadBoolean();
+
+        return frame;
+    }
+
+    private static void ReadSelfShadow(BinaryReader reader, PmmSelfShadow selfShadow)
+    {
+        Current = new("SelfShadow", null, $"The section of self shadow config and frames");
+        selfShadow.EnableSelfShadow = reader.ReadBoolean();
+        selfShadow.ShadowRange = reader.ReadSingle();
+
+        selfShadow.Frames.Add(ReadSelfShadowFrame(reader, true));
+        var frameCount = reader.ReadInt32();
+        for (int i = 0; i < frameCount; i++)
+        {
+            selfShadow.Frames.Add(ReadSelfShadowFrame(reader));
+        }
+    }
+    private static PmmSelfShadowFrame ReadSelfShadowFrame(BinaryReader reader, bool isInitial = false)
+    {
+        // リストの添え字で管理できるため不要なフレームインデックスを破棄
+        if (!isInitial) _ = reader.ReadInt32();
+
+        var frame = new PmmSelfShadowFrame();
+        frame.Frame = reader.ReadInt32();
+        // 所属が確実にわかるので pre/next ID から探索してやる必要性がないため破棄
+        _ = reader.ReadInt32();
+        _ = reader.ReadInt32();
+
+        frame.ShadowMode = (SelfShadow)reader.ReadByte();
+        frame.ShadowRange = reader.ReadSingle();
+
+        frame.IsSelected = reader.ReadBoolean();
+
+        return frame;
+    }
+
+    private static void ReadPhysics(BinaryReader reader, PmmPhysics physics)
+    {
+        Current = new("Physics", null, $"The section of physics config and gravity frames");
+        physics.CalculationMode = (PmmPhysics.PhysicsMode)reader.ReadByte();
+        physics.CurrentGravity.Acceleration = reader.ReadSingle();
+        var currentNoiseAmount = reader.ReadInt32();
+        physics.CurrentGravity.Direction = reader.ReadVector3();
+        var enableNoise = reader.ReadBoolean();
+        physics.CurrentGravity.Noize = enableNoise ? currentNoiseAmount : null;
+
+        physics.GravityFrames.Add(ReadGravityFrame(reader, true));
+        var gravityFrameCount = reader.ReadInt32();
+        for (int i = 0; i < gravityFrameCount; i++)
+        {
+            physics.GravityFrames.Add(ReadGravityFrame(reader));
+        }
+    }
+    private static PmmGravityFrame ReadGravityFrame(BinaryReader reader, bool isInitial = false)
+    {
+        // リストの添え字で管理できるため不要なフレームインデックスを破棄
+        if (!isInitial) _ = reader.ReadInt32();
+
+        var frame = new PmmGravityFrame();
+        frame.Frame = reader.ReadInt32();
+        // 所属が確実にわかるので pre/next ID から探索してやる必要性がないため破棄
+        _ = reader.ReadInt32();
+        _ = reader.ReadInt32();
+
+        var enableNoise = reader.ReadBoolean();
+        var noiseAmount = reader.ReadInt32();
+        frame.Noize = enableNoise ? noiseAmount : null;
+        frame.Acceleration = reader.ReadSingle();
+        frame.Direction = reader.ReadVector3();
+
+        frame.IsSelected = reader.ReadBoolean();
+
+        return frame;
+    }
+
+    private static (PmmAccessory Accessory, byte RenderOrder) ReadAccessory(BinaryReader reader, PolygonMovieMaker pmm)
+    {
+        byte renderOrder;
+
+        // リストの添字で管理するため Index は破棄
+        _ = reader.ReadByte();
+        var name = reader.ReadString(100, Encoding.ShiftJIS, '\0');
+        var path = reader.ReadString(256, Encoding.ShiftJIS, '\0');
+        var acs = new PmmAccessory(name, path);
+
+        renderOrder = reader.ReadByte();
+
+        acs.Frames.Add(ReadAccessoryFrame(reader, pmm, true));
+
+        var accessoryCount = reader.ReadInt32();
+        for (int i = 0; i < accessoryCount; i++)
+        {
+            acs.Frames.Add(ReadAccessoryFrame(reader, pmm));
+        }
+
+        acs.Current.TransAndVisible = reader.ReadByte();
+        int parentModelIndex = reader.ReadInt32();
+        int parentBoneIndex = reader.ReadInt32();
+        acs.Current.ParentModel = parentModelIndex < 0 ? null : pmm.Models[parentModelIndex];
+        acs.Current.ParentBone = acs.Current.ParentModel?.Bones[parentBoneIndex];
+
+        acs.Current.Position = reader.ReadVector3();
+        acs.Current.Rotation = reader.ReadVector3();
+        acs.Current.Scale = reader.ReadSingle();
+
+        acs.Current.EnableShadow = reader.ReadBoolean();
+
+        acs.EnableAlphaBlend = reader.ReadBoolean();
+
+        return (acs, renderOrder);
+    }
+    private static PmmAccessoryFrame ReadAccessoryFrame(BinaryReader reader, PolygonMovieMaker pmm, bool isInitial = false)
+    {
+        // リストの添え字で管理できるため不要なフレームインデックスを破棄
+        if (!isInitial) _ = reader.ReadInt32();
+
+        var frame = new PmmAccessoryFrame();
+
+        frame.Frame = reader.ReadInt32();
+        // 所属が確実にわかるので pre/next ID から探索してやる必要性がないため破棄
+        _ = reader.ReadInt32();
+        _ = reader.ReadInt32();
+
+        frame.TransAndVisible = reader.ReadByte();
+
+        int parentModelIndex = reader.ReadInt32();
+        int parentBoneIndex = reader.ReadInt32();
+
+        frame.ParentModel = parentModelIndex < 0 ? null : pmm.Models[parentModelIndex];
+        frame.ParentBone = frame.ParentModel?.Bones[parentBoneIndex];
+
+        frame.Position = reader.ReadVector3();
+        frame.Rotation = reader.ReadVector3();
+        frame.Scale = reader.ReadSingle();
+
+        frame.EnableShadow = reader.ReadBoolean();
+        frame.IsSelected = reader.ReadBoolean();
+
+        return frame;
+    }
+
+    private static (PmmModel Model, byte RenderOrder, byte CalculateOrder) ReadModel(BinaryReader reader)
+    {
+        var model = new PmmModel();
+        byte renderOrder;
+
+        // モデルのインデックス
+        // 手動で書き換える理由はなく、書込時に自動計算できるので破棄
+        _ = reader.ReadByte();
+
+        model.Name = reader.ReadString();
+        model.NameEn = reader.ReadString();
+        model.Path = reader.ReadString(256, Encoding.ShiftJIS, '\0');
+
+        // キーフレームエディタの行数
+        // 表示枠数から求まるので破棄
+        _ = reader.ReadByte();
+
+        var boneCount = reader.ReadInt32();
+        for (int i = 0; i < boneCount; i++)
+        {
+            Current = new("Bone", i, $"The section of {DataSection.GetOrdinal(i)} bone data in {model.Name}.");
+            model.Bones.Add(new PmmBone(reader.ReadString()));
+        }
+
+        var morphCount = reader.ReadInt32();
+        for (int i = 0; i < morphCount; i++)
+        {
+            Current = new("Morph", i, $"The section of {DataSection.GetOrdinal(i)} morph data in {model.Name}.");
+            model.Morphs.Add(new PmmMorph(reader.ReadString()));
+        }
+
+        Current = new("IK", null, $"The section of IK bone indices.");
+        var ikCount = reader.ReadInt32();
+        var ikIndices = Enumerable.Range(0, ikCount).Select(_ => reader.ReadInt32()).ToArray();
+        foreach (var i in ikIndices)
+        {
+            model.Bones[i].IsIK = true;
+        }
+
+        Current = new("ParentableBone", null, $"The section of outer parentable bone indices.");
+        var parentableBoneCount = reader.ReadInt32();
+        var parentableIndices = Enumerable.Range(0, parentableBoneCount).Select(_ => reader.ReadInt32()).ToArray();
+        // なぜか最初に -1 が入っているのでそれは飛ばす
+        foreach (var i in parentableIndices.Skip(1))
+        {
+            model.Bones[i].CanBecomeOuterParent = true;
+        }
+
+        Current = new("RenderOrder", null, $"The section of render order.");
+        renderOrder = reader.ReadByte();
+        model.CurrentConfig.Visible = reader.ReadBoolean();
+
+        Current = new("SelectedItem", null, $"The section of selected bone/morph index.");
+        int selectedBoneIndex = reader.ReadInt32();
+        int selectedBrowMorphIndex = reader.ReadInt32();
+        int selectedEyeMorphIndex = reader.ReadInt32();
+        int selectedLipMorphIndex = reader.ReadInt32();
+        int selectedOtherMorphIndex = reader.ReadInt32();
+
+        if (selectedBoneIndex >= 0)
+            model.SelectedBone = model.Bones[selectedBoneIndex];
+        if (selectedBrowMorphIndex >= 0)
+            model.SelectedBrowMorph = model.Morphs[selectedBrowMorphIndex];
+        if (selectedEyeMorphIndex >= 0)
+            model.SelectedEyeMorph = model.Morphs[selectedEyeMorphIndex];
+        if (selectedLipMorphIndex >= 0)
+            model.SelectedLipMorph = model.Morphs[selectedLipMorphIndex];
+        if (selectedOtherMorphIndex >= 0)
+            model.SelectedOtherMorph = model.Morphs[selectedOtherMorphIndex];
+
+        var nodeCount = reader.ReadByte();
+        for (int i = 0; i < nodeCount; i++)
+        {
+            Current = new("Morph", i, $"The section of {DataSection.GetOrdinal(i)} node.");
+            model.Nodes.Add(new() { DoesOpen = reader.ReadBoolean() });
+        }
+
+        Current = new("SpecificEditorState", null, $"The section of specific editor state.");
+        model.SpecificEditorState.VerticalScrollState = reader.ReadInt32();
+        model.SpecificEditorState.LastFrame = reader.ReadInt32();
+
+        // 初期ボーンフレームの読込
+        var boneFrameDictionary = new Dictionary<PmmBone, int?>();
+        foreach (var (bone, i) in model.Bones.Select((b, i) => (b, i)))
+        {
+            Current = new("InitialBoneFrame", i, $"The section of {DataSection.GetOrdinal(i)} initial bone frame data in {bone.Name}.");
+            (var frame, _, var nextId, _) = ReadBoneFrame(reader, true);
+            bone.Frames.Add(frame);
+            boneFrameDictionary.Add(bone, nextId);
+        }
+
+        // ボーンフレームの読込
+        ReadFramesThatRequireResolving(
+            reader,
+            static reader => ReadBoneFrame(reader),
+            boneFrameDictionary,
+            static (element, frame) => element?.Frames.Add((PmmBoneFrame)frame)
+        );
+
+        // 初期モーフフレームの読込
+        var morphFrameDictionary = new Dictionary<PmmMorph, int?>();
+        foreach (var (morph, i) in model.Morphs.Select((m, i) => (m, i)))
+        {
+            Current = new("InitialMorphFrame", i, $"The section of {DataSection.GetOrdinal(i)} initial morph frame data in {morph.Name}.");
+            (var frame, _, var nextId, _) = ReadMorphFrame(reader, true);
+            morph.Frames.Add(frame);
+            morphFrameDictionary.Add(morph, nextId);
+        }
+
+        // モーフフレームの読込
+        ReadFramesThatRequireResolving(
+            reader,
+            static reader => ReadMorphFrame(reader),
+            morphFrameDictionary,
+            static (element, frame) => element?.Frames.Add((PmmMorphFrame)frame)
+        );
+
+        Current = new("InitialConfigFrame", null, $"The section of initial config frame in {model.Name}.");
+        model.ConfigFrames.Add(ReadConfigFrame(reader, model, ikIndices, parentableIndices, true));
+
+        var configFrameCount = reader.ReadInt32();
+        for (int i = 0; i < configFrameCount; i++)
+        {
+            Current = new("ConfigFrame", i, $"The section of {DataSection.GetOrdinal(i)} initial config frame data in {model.Name}.");
+            model.ConfigFrames.Add(ReadConfigFrame(reader, model, ikIndices, parentableIndices));
+        }
+
+        foreach (var (bone, i) in model.Bones.Select((b, i) => (b, i)))
+        {
+            Current = new("CurrentBone", i, $"The section of {DataSection.GetOrdinal(i)} current {bone.Name} bone state");
+            bone.Current.Movement = reader.ReadVector3();
+            bone.Current.Rotation = reader.ReadQuaternion();
+            bone.IsCommitted = reader.ReadBoolean();
+            bone.Current.EnablePhysic = reader.ReadBoolean();
+            bone.IsSelected = reader.ReadBoolean();
+        }
+
+        foreach (var (morph, i) in model.Morphs.Select((m, i) => (m, i)))
+        {
+            Current = new("CurrentMorph", i, $"The section of {DataSection.GetOrdinal(i)} current {morph.Name} morph state");
+            morph.Current.Weight = reader.ReadSingle();
+        }
+
+        foreach (var i in ikIndices)
+        {
+            Current = new("CurrentConfig", i, $"The section of {DataSection.GetOrdinal(i)} current config state (IK state of {model.Bones[i].Name}).");
+            model.CurrentConfig.EnableIK.Add(model.Bones[i], reader.ReadBoolean());
+        }
+
+        OuterParentRelationCurrent.Add(model.CurrentConfig, new());
+        foreach (var i in parentableIndices)
+        {
+            Current = new("OuterParentState", i, $"The section of {DataSection.GetOrdinal(i)} outer parent state");
+            var op = new PmmOuterParentState();
+            op.StartFrame = reader.ReadInt32();
+            op.EndFrame = reader.ReadInt32();
+            (int, int) relation = (reader.ReadInt32(), reader.ReadInt32());
+            if (i >= 0)
+            {
+                OuterParentRelationCurrent[model.CurrentConfig].Add(model.Bones[i], relation);
+                model.CurrentConfig.OuterParent.Add(model.Bones[i], op);
             }
         }
 
-        private static void ReadCamera(BinaryReader reader, PmmCamera camera)
+        Current = new("DrawInfo", null, "The section of draw information.");
+        model.EnableAlphaBlend = reader.ReadBoolean();
+        model.EdgeWidth = reader.ReadSingle();
+        model.EnableSelfShadow = reader.ReadBoolean();
+        var calculateOrder = reader.ReadByte();
+
+        return (model, renderOrder, calculateOrder);
+    }
+    private static PmmModelConfigFrame ReadConfigFrame(BinaryReader reader, PmmModel model, IEnumerable<int> ikIndices, IEnumerable<int> parentableIndices, bool isInitial = false)
+    {
+        // リストの添え字で管理できるため不要なフレームインデックスを破棄
+        if (!isInitial) _ = reader.ReadInt32();
+
+        var frame = new PmmModelConfigFrame();
+        OuterParentRelation.Add(frame, new());
+
+        frame.Frame = reader.ReadInt32();
+
+        // ConfigFrame は属するモデルが確実にわかるので pre/next ID から探索してやる必要性がない
+        // なので pre/next ID は破棄する
+        _ = reader.ReadInt32();
+        _ = reader.ReadInt32();
+
+        frame.Visible = reader.ReadBoolean();
+
+        foreach (var i in ikIndices)
         {
-            camera.InitialFrame = ReadCameraFrame(reader, null);
+            frame.EnableIK[model.Bones[i]] = reader.ReadBoolean();
+        }
 
-            var cameraCount = reader.ReadInt32();
-            for (int i = 0; i < cameraCount; i++)
-                camera.Frames.Add(ReadCameraFrame(reader, reader.ReadInt32()));
+        foreach (var i in parentableIndices)
+        {
+            (int, int) relation = (reader.ReadInt32(), reader.ReadInt32());
+            // parentableIndices 最初の要素は -1 なので飛ばす
+            if (i >= 0)
+                OuterParentRelation[frame].Add(model.Bones[i], relation);
+        }
 
-            camera.Uncomitted.EyePosition = reader.ReadVector3();
-            camera.Uncomitted.TargetPosition = reader.ReadVector3();
-            camera.Uncomitted.Rotation = reader.ReadVector3();
-            camera.Uncomitted.EnablePerspective = reader.ReadBoolean();
+        frame.IsSelected = reader.ReadBoolean();
 
-            PmmCameraFrame ReadCameraFrame(BinaryReader reader, int? index)
+        return frame;
+    }
+    private static (PmmMorphFrame Frame, int PreviousFrameIndex, int NextFrameIndex, int? FrameIndex) ReadMorphFrame(BinaryReader reader, bool isInitial = false)
+    {
+        int? id = isInitial ? null : reader.ReadInt32();
+
+        var frame = new PmmMorphFrame();
+
+        frame.Frame = reader.ReadInt32();
+
+        var preID = reader.ReadInt32();
+        var nextId = reader.ReadInt32();
+
+        frame.Weight = reader.ReadSingle();
+        frame.IsSelected = reader.ReadBoolean();
+
+        return (frame, preID, nextId, id);
+    }
+    private static (PmmBoneFrame Frame, int PreviousFrameIndex, int NextFrameIndex, int? FrameIndex) ReadBoneFrame(BinaryReader reader, bool isInitial = false)
+    {
+        int? id = isInitial ? null : reader.ReadInt32();
+
+        var frame = new PmmBoneFrame();
+
+        frame.Frame = reader.ReadInt32();
+        var previousFrameIndex = reader.ReadInt32();
+        var nextFrameIndex = reader.ReadInt32();
+
+        frame.InterpolationCurves[InterpolationItem.XPosition].FromBytes(reader.ReadBytes(4));
+        frame.InterpolationCurves[InterpolationItem.YPosition].FromBytes(reader.ReadBytes(4));
+        frame.InterpolationCurves[InterpolationItem.ZPosition].FromBytes(reader.ReadBytes(4));
+        frame.InterpolationCurves[InterpolationItem.Rotation].FromBytes(reader.ReadBytes(4));
+
+        frame.Movement = reader.ReadVector3();
+        frame.Rotation = reader.ReadQuaternion();
+        frame.IsSelected = reader.ReadBoolean();
+        frame.EnablePhysic = !reader.ReadBoolean();
+
+        return (frame, previousFrameIndex, nextFrameIndex, id);
+    }
+
+    /// <summary>
+    /// ボーン/モーフの初期以外のフレームを読み込む
+    /// <para>フレームの属する要素が前後フレームIDによって管理されているので、各要素内のフレームコレクションに投入するための解決処理を実施する</para>
+    /// </summary>
+    /// <param name="reader">バイナリ読込クラス</param>
+    /// <param name="readElementFrame">フレーム読込メソッドの呼び出し関数</param>
+    /// <param name="elementNextFrameDictionary">要素名とそれに対応する次フレームIDの辞書</param>
+    /// <param name="addFrame">所属要素にフレームを追加する関数</param>
+    private static void ReadFramesThatRequireResolving<T>(
+        BinaryReader reader,
+        Func<BinaryReader, (IPmmFrame? Frame, int PreviousFrameIndex, int NextFrameIndex, int? FrameIndex)> readElementFrame,
+        Dictionary<T, int?> elementNextFrameDictionary,
+        Action<T, IPmmFrame> addFrame)
+     where T : IPmmModelElement
+    {
+        var elementFrameCount = reader.ReadInt32();
+        // 自前定義 Select の1つ目のラムダ式は各ループにおける通常の Select 処理の直前にやる処理
+        var elementFrames = Enumerable.Range(0, elementFrameCount).Select(
+            i => Current = new("Frame", i, $"The section of {DataSection.GetOrdinal(i)} {typeof(T).Name} frame data."),
+            i => readElementFrame(reader)
+        ).ToArray();
+
+        Current = new("ResolveFrames", null, $"This section resolves which {typeof(T).Name} the frame belongs to; the PMM file stores this information by frame ID before and after.");
+
+        ResolveNextFrames(
+            addFrame,
+            elementFrames,
+            elementNextFrameDictionary.Select(p => (Element: p.Key, NextFrameIndex: p.Value == 0 ? null : p.Value)).ToArray()
+        );
+        
+        static void ResolveNextFrames(Action<T, IPmmFrame> addFrame, (IPmmFrame? Frame, int PreviousFrameIndex, int NextFrameIndex, int? FrameIndex)[] elementFrames, (T Element, int? NextFrameIndex)[] nextFramesOfElements)
+        {
+            if (!nextFramesOfElements.Any())
+                return;
+
+            var nextFrames = nextFramesOfElements.Where(p => p.NextFrameIndex.HasValue)
+                                                 .Select(p => (
+                                                    Element: p.Element,
+                                                    NextFrame: elementFrames.FirstOrDefault(f => f.FrameIndex == p.NextFrameIndex, (null, -1, -1, null))
+                                                 ))
+                                                 .ToArray();
+
+            foreach (var (element, nextFrame) in nextFrames.Where(f => f.NextFrame.Frame is not null))
             {
-                var frame = new PmmCameraFrame();
-
-                frame.Index = index;
-
-                frame.Frame = reader.ReadInt32();
-                frame.PreviousFrameIndex = reader.ReadInt32();
-                frame.NextFrameIndex = reader.ReadInt32();
-
-                frame.Distance = reader.ReadSingle();
-                frame.EyePosition = reader.ReadVector3();
-                frame.Rotation = reader.ReadVector3();
-
-                frame.FollowingModelIndex = reader.ReadInt32();
-                frame.FollowingBoneIndex = reader.ReadInt32();
-
-                frame.InterpolationCurces[InterpolationItem.XPosition].FromBytes(reader.ReadBytes(4));
-                frame.InterpolationCurces[InterpolationItem.YPosition].FromBytes(reader.ReadBytes(4));
-                frame.InterpolationCurces[InterpolationItem.ZPosition].FromBytes(reader.ReadBytes(4));
-                frame.InterpolationCurces[InterpolationItem.Rotation].FromBytes(reader.ReadBytes(4));
-                frame.InterpolationCurces[InterpolationItem.Distance].FromBytes(reader.ReadBytes(4));
-                frame.InterpolationCurces[InterpolationItem.ViewAngle].FromBytes(reader.ReadBytes(4));
-
-                frame.EnablePerspective = reader.ReadBoolean();
-                frame.ViewAngle = reader.ReadInt32();
-
-                frame.IsSelected = reader.ReadBoolean();
-
-                return frame;
+                addFrame(element, nextFrame.Frame);
             }
-        }
 
-        private static void ReadLight(BinaryReader reader, PmmLight light)
-        {
-            light.InitialFrame = ReadLightFrame(reader, null);
-
-            var lightCount = reader.ReadInt32();
-            for (int i = 0; i < lightCount; i++)
-                light.Frames.Add(ReadLightFrame(reader, reader.ReadInt32()));
-
-            light.Uncomitted.Color = reader.ReadSingleRGB();
-            light.Uncomitted.Position = reader.ReadVector3();
-
-            PmmLightFrame ReadLightFrame(BinaryReader reader, int? index) => new()
-            {
-                Index = index,
-
-                Frame = reader.ReadInt32(),
-                PreviousFrameIndex = reader.ReadInt32(),
-                NextFrameIndex = reader.ReadInt32(),
-
-                Color = reader.ReadSingleRGB(),
-                Position = reader.ReadVector3(),
-
-                IsSelected = reader.ReadBoolean(),
-            };
-        }
-
-        private static void ReadAccessoryState(BinaryReader reader, PmmEditorState editorState)
-        {
-            editorState.SelectedAccessoryIndex = reader.ReadByte();
-            editorState.VerticalScrollOfAccessoryRow = reader.ReadInt32();
-        }
-
-        private static PmmAccessory ReadAccessory(BinaryReader reader)
-        {
-            var acs = new PmmAccessory();
-
-            acs.Index = reader.ReadByte();
-
-            acs.Name = reader.ReadString(100, Encoding.ShiftJIS, '\0');
-            acs.Path = reader.ReadString(256, Encoding.ShiftJIS, '\0');
-
-            acs.RenderOrder = reader.ReadByte();
-
-            acs.InitialFrame = ReadAccessoryFrame(reader, null);
-
-            var accessoryCount = reader.ReadInt32();
-            for (int i = 0; i < accessoryCount; i++)
-                acs.Frames.Add(ReadAccessoryFrame(reader, reader.ReadInt32()));
-
-            acs.Uncomitted = ReadTmp(reader);
-
-            acs.EnableAlphaBlend = reader.ReadBoolean();
-
-            return acs;
-
-            PmmAccessoryFrame ReadAccessoryFrame(BinaryReader reader, int? index) => new()
-            {
-                Index = index,
-
-                Frame = reader.ReadInt32(),
-                PreviousFrameIndex = reader.ReadInt32(),
-                NextFrameIndex = reader.ReadInt32(),
-
-                OpacityAndVisible = reader.ReadByte(),
-
-                ParentModelIndex = reader.ReadInt32(),
-                ParentBoneIndex = reader.ReadInt32(),
-
-                Position = reader.ReadVector3(),
-                Rotation = reader.ReadVector3(),
-                Scale = reader.ReadSingle(),
-
-                EnableShadow = reader.ReadBoolean(),
-
-                IsSelected = reader.ReadBoolean(),
-            };
-
-            TemporaryAccessoryEditState ReadTmp(BinaryReader reader) => new()
-            {
-                OpacityAndVisible = reader.ReadByte(),
-
-                ParentModelIndex = reader.ReadInt32(),
-                ParentBoneIndex = reader.ReadInt32(),
-
-                Position = reader.ReadVector3(),
-                Rotation = reader.ReadVector3(),
-                Scale = reader.ReadSingle(),
-
-                EnableShadow = reader.ReadBoolean(),
-            };
-        }
-
-        private static void ReadFrameState(BinaryReader reader, PmmEditorState editorState)
-        {
-            editorState.CurrentFrame = reader.ReadInt32();
-            editorState.HorizontalScroll = reader.ReadInt32();
-            editorState.HorizontalScrollLength = reader.ReadInt32();
-            editorState.SelectedBoneOperation = (PmmEditorState.BoneOperation)reader.ReadInt32();
-        }
-
-        private static void ReadPlayConfig(BinaryReader reader, PmmPlayConfig playConfig)
-        {
-            playConfig.CameraTrackingTarget = (PmmPlayConfig.TrackingTarget)reader.ReadByte();
-
-            playConfig.IsRepeat = reader.ReadBoolean();
-            playConfig.IsMoveCurrentFrameToStopFrame = reader.ReadBoolean();
-            playConfig.IsStartFromCurrentFrame = reader.ReadBoolean();
-
-            playConfig.PlayStartFrame = reader.ReadInt32();
-            playConfig.PlayStopFrame = reader.ReadInt32();
-        }
-
-        private static void ReadMediaConfig(BinaryReader reader, PmmMediaConfig mediaConfig)
-        {
-            mediaConfig.EnableAudio = reader.ReadBoolean();
-            mediaConfig.AudioPath = reader.ReadString(256, Encoding.ShiftJIS, '\0');
-
-            mediaConfig.BackgroundVideoOffset = new(reader.ReadInt32(), reader.ReadInt32());
-            mediaConfig.BackgroundVideoScale = reader.ReadSingle();
-            mediaConfig.BackgroundVideoPath = reader.ReadString(256, Encoding.ShiftJIS, '\0');
-            mediaConfig.EnableBackgroundVideo = reader.ReadInt32() == 0b01000000;
-
-            mediaConfig.BackgroundImageOffset = new(reader.ReadInt32(), reader.ReadInt32());
-            mediaConfig.BackgroundImageScale = reader.ReadSingle();
-            mediaConfig.BackgroundImagePath = reader.ReadString(256, Encoding.ShiftJIS, '\0');
-            mediaConfig.EnableBackgroundImage = reader.ReadBoolean();
-        }
-
-        private static void ReadDrawConfig(BinaryReader reader, PmmViewConfig drawConfig)
-        {
-            drawConfig.IsShowInfomation = reader.ReadBoolean();
-            drawConfig.IsShowAxis = reader.ReadBoolean();
-            drawConfig.IsShowGrandShadow = reader.ReadBoolean();
-
-            drawConfig.FPSLimit = reader.ReadSingle();
-            drawConfig.ScreenCaptureSetting = (PmmViewConfig.ScreenCaptureMode)reader.ReadInt32();
-            drawConfig.AccessoryModelThreshold = reader.ReadInt32();
-
-            drawConfig.GroundShadowBrightness = reader.ReadSingle();
-            drawConfig.EnableTransparentGroundShadow = reader.ReadBoolean();
-
-            drawConfig.PhysicsSetting = (PmmViewConfig.PhysicsMode)reader.ReadByte();
-        }
-
-        private static void ReadGravity(BinaryReader reader, PmmGravity gravity)
-        {
-            gravity.Acceleration = reader.ReadSingle();
-            gravity.NoizeAmount = reader.ReadInt32();
-            gravity.Direction = reader.ReadVector3();
-            gravity.EnableNoize = reader.ReadBoolean();
-
-            gravity.InitialFrame = ReadGravityFrame(reader, null);
-
-            var frameCount = reader.ReadInt32();
-            for (int i = 0; i < frameCount; i++)
-                gravity.Frames.Add(ReadGravityFrame(reader, reader.ReadInt32()));
-
-            PmmGravityFrame ReadGravityFrame(BinaryReader reader, int? index) => new()
-            {
-                Index = index,
-
-                Frame = reader.ReadInt32(),
-                PreviousFrameIndex = reader.ReadInt32(),
-                NextFrameIndex = reader.ReadInt32(),
-
-                EnableNoize = reader.ReadBoolean(),
-                NoizeAmount = reader.ReadInt32(),
-                Acceleration = reader.ReadSingle(),
-                Direction = reader.ReadVector3(),
-
-                IsSelected = reader.ReadBoolean(),
-            };
-        }
-
-        private static void ReadSelfShadow(BinaryReader reader, PmmSelfShadow selfShadow)
-        {
-            selfShadow.EnableSelfShadow = reader.ReadBoolean();
-            selfShadow.ShadowLimit = reader.ReadSingle();
-
-            selfShadow.InitialFrame = ReadSelfShadowFrame(reader, null);
-
-            var selfShadowCount = reader.ReadInt32();
-            for (int i = 0; i < selfShadowCount; i++)
-                selfShadow.Frames.Add(ReadSelfShadowFrame(reader, reader.ReadInt32()));
-
-            PmmSelfShadowFrame ReadSelfShadowFrame(BinaryReader reader, int? index) => new()
-            {
-                Index = index,
-
-                Frame = reader.ReadInt32(),
-                PreviousFrameIndex = reader.ReadInt32(),
-                NextFrameIndex = reader.ReadInt32(),
-
-                ShadowMode = (PmmSelfShadowFrame.Shadow)reader.ReadByte(),
-                ShadowRange = reader.ReadSingle(),
-
-                IsSelected = reader.ReadBoolean(),
-            };
-        }
-
-        private static void ReadColorConfig(BinaryReader reader, PmmViewConfig drawConfig)
-        {
-            drawConfig.EdgeColor = System.Drawing.Color.FromArgb(reader.ReadInt32(), reader.ReadInt32(), reader.ReadInt32());
-            drawConfig.IsBackgroundBlack = reader.ReadBoolean();
-        }
-
-        private static void ReadUncomittedFollowingState(BinaryReader reader, PmmCamera camera)
-        {
-            camera.UncomittedFollowing.ModelIndex = reader.ReadInt32();
-            camera.UncomittedFollowing.BoneIndex = reader.ReadInt32();
-        }
-
-        private static void ReadViewFollowing(BinaryReader reader, PmmViewConfig viewConfig)
-        {
-            viewConfig.IsViewFollowCamera = reader.ReadBoolean();
-        }
-
-        private static void ReadGroundPhysics(BinaryReader reader, PmmViewConfig drawConfig)
-        {
-            drawConfig.EnableGroundPhysics = reader.ReadBoolean();
-        }
-
-        private static void ReadFrameLocation(BinaryReader reader, PmmViewConfig viewConfig)
-        {
-            viewConfig.FrameLocation = reader.ReadInt32();
-        }
-
-        private static (PmmModel Model, int Target) ReadRangeSelectionTargetIndex(BinaryReader reader, PolygonMovieMaker pmm)
-        {
-            return (pmm.Models[reader.ReadByte()], reader.ReadInt32());
+            ResolveNextFrames(
+                addFrame,
+                elementFrames,
+                nextFrames.Select<(T Element, (IPmmFrame? Frame, int PreviousFrameIndex, int NextFrameIndex, int? FrameIndex) NextFrame), (T Element, int? NextFrameIndex)>
+                                 (p => (
+                                    p.Element,
+                                    p.NextFrame switch
+                                    {
+                                        { Frame: null } => null,
+                                        { NextFrameIndex: 0 } => null,
+                                        _ => p.NextFrame.NextFrameIndex
+                                    }
+                                 )
+                          ).Where(p => p.NextFrameIndex.HasValue).ToArray()
+            );
         }
     }
 }
