@@ -1,49 +1,10 @@
-﻿using MikuMikuMethods.PMM.Frame;
-using System.Collections.Generic;
-using System.Numerics;
+﻿using MikuMikuMethods.Pmm.ElementState;
+using MikuMikuMethods.Pmm.Frame;
 
-namespace MikuMikuMethods.PMM
+namespace MikuMikuMethods.Pmm;
+
+public class PmmLight
 {
-    /// <summary>
-    /// PMMの照明情報
-    /// </summary>
-    public class PmmLight
-    {
-        /// <summary>
-        /// 初期位置の照明フレーム
-        /// </summary>
-        public PmmLightFrame InitialFrame { get; set; }
-        /// <summary>
-        /// 照明のキーフレーム
-        /// </summary>
-        public List<PmmLightFrame> Frames { get; init; }
-        /// <summary>
-        /// 未確定編集状態
-        /// </summary>
-        public TemporaryLightEditState Uncomitted { get; set; }
-
-        /// <summary>
-        /// コンストラクタ
-        /// </summary>
-        public PmmLight()
-        {
-            Frames = new();
-            Uncomitted = new();
-        }
-    }
-
-    /// <summary>
-    /// 未確定のライト編集状態
-    /// </summary>
-    public class TemporaryLightEditState
-    {
-        /// <summary>
-        /// 色(RGBのみ使用)
-        /// </summary>
-        public ColorF Color { get; set; }
-        /// <summary>
-        /// 位置
-        /// </summary>
-        public Vector3 Position { get; set; }
-    }
+    public List<PmmLightFrame> Frames { get; } = new();
+    public PmmLightState Current { get; } = new();
 }
