@@ -1,29 +1,17 @@
-﻿using MikuMikuMethods.PMM.Frame;
-using System.Collections.Generic;
+﻿using MikuMikuMethods.Pmm.Frame;
 
-namespace MikuMikuMethods.PMM
+namespace MikuMikuMethods.Pmm;
+
+public class PmmSelfShadow
 {
     /// <summary>
-    /// セルフ影情報
+    /// セルフ影のOn/Off
     /// </summary>
-    public class PmmSelfShadow
-    {
-        /// <summary>
-        /// セルフ影のOn/Off
-        /// </summary>
-        public bool EnableSelfShadow { get; set; }
-        /// <summary>
-        /// 影範囲
-        /// </summary>
-        public float ShadowLimit { get; set; }
+    public bool EnableSelfShadow { get; set; } = true;
+    /// <summary>
+    /// 影範囲
+    /// </summary>
+    public float ShadowRange { get; set; } = 8875;
 
-        /// <summary>
-        /// 初期位置のセルフ影フレーム
-        /// </summary>
-        public PmmSelfShadowFrame InitialFrame { get; set; }
-        /// <summary>
-        /// セルフ影のキーフレーム
-        /// </summary>
-        public List<PmmSelfShadowFrame> Frames { get; init; } = new();
-    }
+    public List<PmmSelfShadowFrame> Frames { get; } = new();
 }
