@@ -1,9 +1,8 @@
-﻿namespace MikuMikuMethods.PMX
+﻿namespace MikuMikuMethods.Pmx;
+
+/// <summary>
+/// モーフ内要素のインターフェイス
+/// </summary>
+public interface IPmxOffset : IPmxData
 {
-    /// <summary>
-    /// モーフ内要素のインターフェイス
-    /// </summary>
-    public interface IPmxOffset : IPmxData
-    {
-    }
 }
