@@ -1,26 +1,32 @@
 ﻿using System.Numerics;
 
-namespace MikuMikuMethods.PMX
+namespace MikuMikuMethods.Pmx;
+
+/// <summary>
+/// ボーンモーフのオフセット
+/// </summary>
+public class PmxOffsetBone : IPmxOffset
 {
     /// <summary>
-    /// ボーンモーフのオフセット
+    /// 対象ボーン
     /// </summary>
-    public class PmxOffsetBone : IPmxOffset
+    public PmxBone Target { get; set; }
+
+    /// <summary>
+    /// 移動量
+    /// </summary>
+    public Vector3 Offset { get; set; }
+    /// <summary>
+    /// 回転量
+    /// </summary>
+    public Quaternion Rotate { get; set; }
+
+    public override string ToString() => $"{Target.Name} : {{{Offset} - {Rotate}}}";
+
+    public PmxOffsetBone(PmxBone target, Vector3 offset = default, Quaternion rotate = default)
     {
-        /// <summary>
-        /// 対象ボーン
-        /// </summary>
-        public PmxBone Target { get; set; }
-
-        /// <summary>
-        /// 移動量
-        /// </summary>
-        public Vector3 Offset { get; set; }
-        /// <summary>
-        /// 回転量
-        /// </summary>
-        public Quaternion Rotate { get; set; }
-
-        public override string ToString() => $"{Target.Name} : {{{Offset} - {Rotate}}}";
+        Target = target;
+        Offset = offset;
+        Rotate = rotate;
     }
 }
