@@ -1,22 +1,27 @@
 ﻿using System.Numerics;
 
-namespace MikuMikuMethods.PMX
+namespace MikuMikuMethods.Pmx;
+
+/// <summary>
+/// 頂点モーフ
+/// </summary>
+public class PmxOffsetVertex : IPmxOffset
 {
     /// <summary>
-    /// 頂点モーフ
+    /// 対象頂点
     /// </summary>
-    public class PmxOffsetVertex : IPmxOffset
+    public PmxVertex Target { get; set; }
+
+    /// <summary>
+    /// 移動量
+    /// </summary>
+    public Vector3 Offset { get; set; }
+
+    public override string ToString() => $"{Target.Position} : {Offset}";
+
+    public PmxOffsetVertex(PmxVertex target, Vector3 offset = default)
     {
-        /// <summary>
-        /// 対象頂点
-        /// </summary>
-        public PmxVertex Target { get; set; }
-
-        /// <summary>
-        /// 移動量
-        /// </summary>
-        public Vector3 Offset { get; set; }
-
-        public override string ToString() => $"{Target.Position} : {Offset}";
+        Target = target;
+        Offset = offset;
     }
 }
