@@ -1,57 +1,55 @@
-﻿using System.Collections.Generic;
-using System.Numerics;
+﻿using System.Numerics;
 
-namespace MikuMikuMethods.PMX
+namespace MikuMikuMethods.Pmx;
+
+/// <summary>
+/// IK情報
+/// </summary>
+public class PmxInverseKinematics
 {
     /// <summary>
-    /// IK情報
+    /// IKのターゲットボーン
     /// </summary>
-    public class PmxInverseKinematics
-    {
-        /// <summary>
-        /// IKのターゲットボーン
-        /// </summary>
-        public PmxBone Target { get; set; }
-        /// <summary>
-        /// ループ回数
-        /// </summary>
-        public int LoopNum { get; set; }
-        /// <summary>
-        /// ループ計算時の制限角度 -> ラジアン
-        /// </summary>
-        public float LimitAngle { get; set; }
-
-        /// <summary>
-        /// リンク
-        /// </summary>
-        public List<PmxIKLink> Links { get; } = new();
-
-        public override string ToString() => $"{Target.Name} - {LoopNum} : {LimitAngle:###.00}";
-    }
+    public PmxBone? Target { get; set; }
+    /// <summary>
+    /// ループ回数
+    /// </summary>
+    public int LoopNum { get; set; }
+    /// <summary>
+    /// ループ計算時の制限角度 -> ラジアン
+    /// </summary>
+    public float LimitAngle { get; set; }
 
     /// <summary>
-    /// IKのリンク情報
+    /// リンク
     /// </summary>
-    public class PmxIKLink
-    {
-        /// <summary>
-        /// 対象ボーン
-        /// </summary>
-        public PmxBone Bone { get; set; }
-        
-        /// <summary>
-        /// 角度制限の有無
-        /// </summary>
-        public bool EnableAngleLimit { get; set; }
-        /// <summary>
-        /// 角度制限の上限
-        /// </summary>
-        public Vector3 UpperLimit { get; set; }
-        /// <summary>
-        /// 角度制限の下限
-        /// </summary>
-        public Vector3 LowerLimit { get; set; }
+    public List<PmxIKLink> Links { get; } = new();
 
-        public override string ToString() => $"Link: {Bone.Name}{(EnableAngleLimit ? "(Limited)" : "")}";
-    }
+    public override string ToString() => $"{Target?.Name} - {LoopNum} : {LimitAngle:###.00}";
+}
+
+/// <summary>
+/// IKのリンク情報
+/// </summary>
+public class PmxIKLink
+{
+    /// <summary>
+    /// 対象ボーン
+    /// </summary>
+    public PmxBone? Bone { get; set; }
+
+    /// <summary>
+    /// 角度制限の有無
+    /// </summary>
+    public bool EnableAngleLimit { get; set; }
+    /// <summary>
+    /// 角度制限の上限
+    /// </summary>
+    public Vector3 UpperLimit { get; set; }
+    /// <summary>
+    /// 角度制限の下限
+    /// </summary>
+    public Vector3 LowerLimit { get; set; }
+
+    public override string ToString() => $"Link: {Bone?.Name}{(EnableAngleLimit ? "(Limited)" : "")}";
 }
