@@ -1,101 +1,100 @@
 ﻿using System.Numerics;
 
-namespace MikuMikuMethods.PMX
+namespace MikuMikuMethods.Pmx;
+
+/// <summary>
+/// ジョイント
+/// </summary>
+public class PmxJoint : IPmxData
 {
     /// <summary>
-    /// ジョイント
+    /// 名前
     /// </summary>
-    public class PmxJoint : IPmxData
+    public string Name { get; set; } = "";
+    /// <summary>
+    /// 英名
+    /// </summary>
+    public string NameEn { get; set; } = "";
+
+    /// <summary>
+    /// ジョイント種別
+    /// </summary>
+    public JointType Type { get; set; }
+
+    /// <summary>
+    /// 関連剛体A
+    /// </summary>
+    public PmxBody? RelationBodyA { get; set; }
+    /// <summary>
+    /// 関連剛体B
+    /// </summary>
+    public PmxBody? RelationBodyB { get; set; }
+
+    /// <summary>
+    /// 位置
+    /// </summary>
+    public Vector3 Position { get; set; }
+    /// <summary>
+    /// 回転
+    /// </summary>
+    public Vector3 Rotation { get; set; }
+
+    /// <summary>
+    /// 移動制限-下限
+    /// </summary>
+    public Vector3 MovingLowerLimit { get; set; }
+    /// <summary>
+    /// 移動制限-上限
+    /// </summary>
+    public Vector3 MovingUpperLimit { get; set; }
+    /// <summary>
+    /// 回転制限-下限
+    /// </summary>
+    public Vector3 RotationLowerLimit { get; set; }
+    /// <summary>
+    /// 回転制限-上限
+    /// </summary>
+    public Vector3 RotationUpperLimit { get; set; }
+
+    /// <summary>
+    /// バネ定数-移動
+    /// </summary>
+    public Vector3 MovingSpringConstants { get; set; }
+    /// <summary>
+    /// バネ定数-回転
+    /// </summary>
+    public Vector3 RotationSpringConstants { get; set; }
+
+    /// <summary>
+    /// ジョイント種別
+    /// </summary>
+    public enum JointType : byte
     {
         /// <summary>
-        /// 名前
+        /// バネ付6DOF
         /// </summary>
-        public string Name { get; set; }
+        SixDofWithSpring,
         /// <summary>
-        /// 英名
+        /// 6DOF
         /// </summary>
-        public string NameEn { get; set; }
-
+        SixDof,
         /// <summary>
-        /// ジョイント種別
+        /// P2P
         /// </summary>
-        public JointType Type { get; set; }
-
+        P2P,
         /// <summary>
-        /// 関連剛体A
+        /// ConeTwist
         /// </summary>
-        public PmxBody RelationBodyA { get; set; }
+        ConeTwist,
         /// <summary>
-        /// 関連剛体B
+        /// Slider
         /// </summary>
-        public PmxBody RelationBodyB { get; set; }
-
+        Slider,
         /// <summary>
-        /// 位置
+        /// Hinge
         /// </summary>
-        public Vector3 Position { get; set; }
-        /// <summary>
-        /// 回転
-        /// </summary>
-        public Vector3 Rotation { get; set; }
-
-        /// <summary>
-        /// 移動制限-下限
-        /// </summary>
-        public Vector3 MovingLowerLimit { get; set; }
-        /// <summary>
-        /// 移動制限-上限
-        /// </summary>
-        public Vector3 MovingUpperLimit { get; set; }
-        /// <summary>
-        /// 回転制限-下限
-        /// </summary>
-        public Vector3 RotationLowerLimit { get; set; }
-        /// <summary>
-        /// 回転制限-上限
-        /// </summary>
-        public Vector3 RotationUpperLimit { get; set; }
-
-        /// <summary>
-        /// バネ定数-移動
-        /// </summary>
-        public Vector3 MovingSpringConstants { get; set; }
-        /// <summary>
-        /// バネ定数-回転
-        /// </summary>
-        public Vector3 RotationSpringConstants { get; set; }
-
-        /// <summary>
-        /// ジョイント種別
-        /// </summary>
-        public enum JointType : byte
-        {
-            /// <summary>
-            /// バネ付6DOF
-            /// </summary>
-            SixDofWithSpring,
-            /// <summary>
-            /// 6DOF
-            /// </summary>
-            SixDof,
-            /// <summary>
-            /// P2P
-            /// </summary>
-            P2P,
-            /// <summary>
-            /// ConeTwist
-            /// </summary>
-            ConeTwist,
-            /// <summary>
-            /// Slider
-            /// </summary>
-            Slider,
-            /// <summary>
-            /// Hinge
-            /// </summary>
-            Hinge
-        }
-
-        public override string ToString() => $"{Name} - {Position}";
+        Hinge
     }
+
+    public override string ToString() => $"{Name} - {Position}";
 }
