@@ -1,18 +1,21 @@
-﻿namespace MikuMikuMethods.VMD
-{
-    /// <summary>
-    /// カメラ系フレームの抽象クラス
-    /// </summary>
-    public abstract class VmdCameraTypeFrame : VmdFrame
-    {
-        /// <summary>
-        /// カメラ系フレームか？
-        /// </summary>
-        public override bool IsCameraType => true;
+﻿namespace MikuMikuMethods.Vmd;
 
-        /// <summary>
-        /// モデル系フレームか？
-        /// </summary>
-        public override bool IsModelType => false;
+/// <summary>
+/// カメラ系フレームの抽象クラス
+/// </summary>
+public abstract class VmdCameraTypeFrame : VmdFrame
+{
+    protected VmdCameraTypeFrame(string name) : base(name)
+    {
     }
+
+    /// <summary>
+    /// カメラ系フレームか？
+    /// </summary>
+    public override bool IsCameraType => true;
+
+    /// <summary>
+    /// モデル系フレームか？
+    /// </summary>
+    public override bool IsModelType => false;
 }
