@@ -1,18 +1,21 @@
-﻿namespace MikuMikuMethods.VMD
-{
-    /// <summary>
-    /// モデル系フレームの抽象クラス
-    /// </summary>
-    public abstract class VmdModelTypeFrame : VmdFrame
-    {
-        /// <summary>
-        /// カメラ系フレームか？
-        /// </summary>
-        public override bool IsCameraType => false;
+﻿namespace MikuMikuMethods.Vmd;
 
-        /// <summary>
-        /// モデル系フレームか？
-        /// </summary>
-        public override bool IsModelType => true;
+/// <summary>
+/// モデル系フレームの抽象クラス
+/// </summary>
+public abstract class VmdModelTypeFrame : VmdFrame
+{
+    protected VmdModelTypeFrame(string name) : base(name)
+    {
     }
+
+    /// <summary>
+    /// カメラ系フレームか？
+    /// </summary>
+    public override bool IsCameraType => false;
+
+    /// <summary>
+    /// モデル系フレームか？
+    /// </summary>
+    public override bool IsModelType => true;
 }
